+++ tests/Command/PeekStatusTest.php
@@ -5,18 +5,40 @@ declare(strict_types=1);
 namespace Phlib\Beanstalk\Command;
 
 use Phlib\Beanstalk\Exception\CommandException;
+use Phlib\Beanstalk\Exception\InvalidArgumentException;
 use Phlib\Beanstalk\Exception\NotFoundException;
 
-class PeekTest extends CommandTestCase
+class PeekStatusTest extends CommandTestCase
 {
     public function testImplementsCommand()
     {
-        static::assertInstanceOf(CommandInterface::class, new Peek(10));
+        static::assertInstanceOf(CommandInterface::class, new PeekStatus('ready'));
     }
 
-    public function testGetCommand()
+    /**
+     * @param mixed $subject
+     * @param string $command
+     * @dataProvider getCommandDataProvider
+     */
+    public function testGetCommand($subject, $command)
     {
-        static::assertEquals('peek 123', (new Peek(123))->getCommand());
+        static::assertEquals($command, (new PeekStatus($subject))->getCommand());
+    }
+
+    public function getCommandDataProvider()
+    {
+        return [
+            ['ready', 'peek-ready'],
+            ['delayed', 'peek-delayed'],
+            ['buried', 'peek-buried'],
+        ];
+    }
+
+    public function testWithInvalidSubject()
+    {
+        $this->expectException(InvalidArgumentException::class);
+
+        new PeekStatus('foo-bar');
     }
 
     public function testSuccessfulCommand()
@@ -32,7 +54,7 @@ class PeekTest extends CommandTestCase
             ->method('read')
             ->willReturnOnConsecutiveCalls("FOUND {$id} 54\r\n", $body . "\r\n");
 
-        static::assertEquals($response, (new Peek(10))->process($this->socket));
+        static::assertEquals($response, (new PeekStatus(PeekStatus::BURIED))->process($this->socket));
     }
 
     public function testNotFoundThrowsException()
@@ -42,7 +64,7 @@ class PeekTest extends CommandTestCase
         $this->socket->expects(static::any())
             ->method('read')
             ->willReturn('NOT_FOUND');
-        (new Peek(10))->process($this->socket);
+        (new PeekStatus(PeekStatus::BURIED))->process($this->socket);
     }
 
     public function testUnknownStatusThrowsException()
@@ -52,6 +74,6 @@ class PeekTest extends CommandTestCase
         $this->socket->expects(static::any())
             ->method('read')
             ->willReturn('UNKNOWN_ERROR');
-        (new Peek(10))->process($this->socket);
+        (new PeekStatus(PeekStatus::BURIED))->process($this->socket);
     }
 }